backend/requirements.txt
# Core dependencies
flask>=3.0.0
flask-cors>=4.0.0
flask-restx>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0

# AI and ML dependencies
numpy>=1.24.0
pandas>=2.0.3
scikit-learn>=1.3.0
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0
torchvision>=0.15.0
transformers>=4.35.0
tokenizers>=0.11.0
sentence-transformers==2.2.2
scipy>=1.11.1
datasets>=1.18.3

# Explainability
shap>=0.42.1
lime>=0.2.0.1

# Async and networking
aiohttp>=3.8.0
asyncio>=3.4.3

# Database and caching
redis>=4.5.0
psycopg2-binary>=2.9.0
supabase>=1.0.0
sqlite3

# Security and encryption
cryptography>=3.4.0
hmac>=3.2
hashlib>=3.2
secrets>=1.0.0

# Utilities
python-dateutil>=2.8.0
backoff>=2.2.1
python-multipart>=0.0.6
pydantic>=2.0.0
typing-extensions>=4.7.0
dataclasses>=0.6
setuptools>=65.0.0
wheel>=0.38.0
pickle5>=0.0.12
concurrent.futures>=3.1.1

# Logging and Monitoring
structlog>=23.1.0
logging>=0.4.9.6
prometheus-client>=0.17.1

# Scheduling
schedule>=1.2.0

# For MaterialsBERT and NLP
sentence-transformers==2.2.2

